+++ src/main/scala/dotty/dokka/model/extras.scala
@@ -2,16 +2,21 @@ package dotty.dokka
 
 import org.jetbrains.dokka.DokkaConfiguration$DokkaSourceSet
 import org.jetbrains.dokka.links._
+import org.jetbrains.dokka.model.{Projection => JProjection}
 import org.jetbrains.dokka.model._
 import collection.JavaConverters._
 import org.jetbrains.dokka.links._
 import org.jetbrains.dokka.model.doc._
 import org.jetbrains.dokka.model.properties._  
 
+
+
 case class IsGiven(givenInstance: Option[Bound]) extends ExtraProperty[Documentable]:
   override def getKey = IsGiven
 
 object IsGiven extends BaseKey[Documentable, IsGiven]
+
+
 
 case class ExtensionInformation(val isGrouped: Boolean)
    
@@ -20,10 +25,14 @@ case class MethodExtension(parametersListSizes: Seq[Int], extensionInfo: Option[
 
 object MethodExtension extends BaseKey[DFunction, MethodExtension]
 
+
+
 case class ParameterExtension(isExtendedSymbol: Boolean, isGrouped: Boolean) extends ExtraProperty[DParameter]:
   override def getKey = ParameterExtension
 
 object ParameterExtension extends BaseKey[DParameter, ParameterExtension]
+
+
 
 enum IsEnumEntry extends ExtraProperty[Documentable]:
   case Val
@@ -33,22 +42,22 @@ enum IsEnumEntry extends ExtraProperty[Documentable]:
 
 object IsEnumEntry extends BaseKey[Documentable, IsEnumEntry]
 
-enum Kind(val name: String){
-  case Class extends Kind("class")
-  case Object extends Kind("object")
-  case Trait extends Kind("trait")
-  case Enum extends Kind("enum")
-}
 
-case class ExtensionGroup(val extendedSymbol: DParameter, val extensions: List[DFunction])
 
 case class EnumExtension(val enumEntries: Seq[Documentable]) extends ExtraProperty[DClass]:
   override def getKey = EnumExtension
 
 object EnumExtension extends BaseKey[DClass, EnumExtension]
 
-case class TastyDocumentableSource(val path: String, val lineNumber: Int) extends DocumentableSource {
-    override def getPath = path
+
+
+case class ExtensionGroup(val extendedSymbol: DParameter, val extensions: List[DFunction])
+
+enum Kind(val name: String){
+  case Class extends Kind("class")
+  case Object extends Kind("object")
+  case Trait extends Kind("trait")
+  case Enum extends Kind("enum")
 }
 
 case class ClasslikeExtension(
@@ -64,6 +73,8 @@ case class ClasslikeExtension(
 
 object ClasslikeExtension extends BaseKey[DClasslike, ClasslikeExtension]
 
+
+
 case class InheritanceInfo(
   val parents: List[Bound],
   val knownChildren: List[DRI]
@@ -72,37 +83,18 @@ case class InheritanceInfo(
 
 object InheritanceInfo extends BaseKey[DClasslike, InheritanceInfo]  
 
+
+
 case class PropertyExtension(kind: "val" | "var" | "type", isAbstract: Boolean) extends ExtraProperty[DProperty]:
   override def getKey = PropertyExtension
 
 object PropertyExtension extends BaseKey[DProperty, PropertyExtension]
 
-class BaseKey[T, V] extends ExtraProperty.Key[T, V]:
-  override def mergeStrategyFor(left: V, right: V): MergeStrategy[T] = 
-    MergeStrategy.Remove.INSTANCE.asInstanceOf[MergeStrategy[T]]
 
-enum ScalaOnlyModifiers(val name: String, val prefix: Boolean) extends ExtraModifiers(name, null):
-  case Sealed extends ScalaOnlyModifiers("sealed", true)
-  case Case extends ScalaOnlyModifiers("case", false)
-  case Implicit extends ScalaOnlyModifiers("implicit", true)
-  case Inline extends ScalaOnlyModifiers("inline", true)
-  case Lazy extends ScalaOnlyModifiers("lazy", true)
-  case Override extends ScalaOnlyModifiers("override", true)
-  case Erased extends ScalaOnlyModifiers("erased", true)
-  case Opaque extends ScalaOnlyModifiers("opaque", true)
-    
-enum ScalaVisibility(val name: String) extends org.jetbrains.dokka.model.Visibility(name, null):
-  case NoModifier extends ScalaVisibility("")
-  case Protected extends ScalaVisibility("protected")
-  case Private extends ScalaVisibility("private")
 
-enum ScalaModifier(val name: String) extends org.jetbrains.dokka.model.Modifier(name, null):
-  case Abstract extends ScalaModifier("abstract")
-  case Final extends ScalaModifier("final")
-  case Empty extends ScalaModifier("")
+case class AnnotationsInfo(val annotations: List[AnnotationsInfo.Annotation]) extends ExtraProperty[Documentable]:
 
-extension (f: DFunction):
-  def isRightAssociative(): Boolean = f.getName.endsWith(":")
-  def getExtendedSymbol(): Option[DParameter] = Option.when(f.get(MethodExtension).extensionInfo.isDefined)(
-    f.getParameters.asScala(if f.isRightAssociative() then f.get(MethodExtension).parametersListSizes(0) else 0)
-  )
+    override def getKey = AnnotationsInfo
+
+object AnnotationsInfo extends BaseKey[Documentable, AnnotationsInfo]:
+    case class Annotation(val inner: List[JProjection])
