+++ src/main/scala/dotty/dokka/transformers/InheritanceInformationTransformer.scala
